+++ pyvela/tests/test_run_analysis.py
@@ -5,7 +5,17 @@ import emcee
 import numpy as np
 
 from pyvela.spnta import SPNTA
+from pyvela import pyvela_script
 from pint.models import get_model_and_toas
+
+prior_str = """
+    {
+        "EFAC": {
+            "distribution": "Uniform",
+            "args": [0.5, 1.5]
+        }
+    }
+"""
 
 
 def test_analysis_NGC6440E_emcee():
@@ -15,19 +25,10 @@ def test_analysis_NGC6440E_emcee():
 
     mp, tp = get_model_and_toas(parfile, timfile, planets=True)
 
-    prior_str = StringIO(
-        """{
-            "EFAC": {
-                "distribution": "Uniform",
-                "args": [0.5, 1.5]
-            }
-        }"""
-    )
-
     spnta = SPNTA.from_pint(
         mp,
         tp,
-        custom_priors=prior_str,
+        custom_priors=StringIO(prior_str),
     )
 
     nwalkers = 3 * spnta.ndim
@@ -50,3 +51,27 @@ def test_analysis_NGC6440E_emcee():
         and pint_model_final[pname].uncertainty_value != 0
         for pname in pint_model_final.free_params
     )
+
+
+def test_script_NGC6440():
+    datadir = os.path.dirname(os.path.realpath(__file__)) + "/datafiles"
+    dataset = "NGC6440E"
+    parfile, timfile = f"{datadir}/{dataset}.par", f"{datadir}/{dataset}.tim"
+    outdir = "_NGC6440E_out"
+
+    prior_file = "__prior.json"
+    with open(prior_file, "w") as pf:
+        print(prior_str, file=pf)
+
+    args = f"{parfile} {timfile} -P {prior_file} -o {outdir}".split()
+
+    pyvela_script.main(args)
+
+    assert os.path.isdir(outdir)
+    assert os.path.isfile(f"{outdir}/summary.json")
+    assert os.path.isfile(f"{outdir}/{prior_file}")
+    assert os.path.isfile(f"{outdir}/param_names.txt")
+    assert os.path.isfile(f"{outdir}/param_units.txt")
+    assert os.path.isfile(f"{outdir}/param_scale_factors.txt")
+    assert os.path.isfile(f"{outdir}/samples_raw.npy")
+    assert os.path.isfile(f"{outdir}/samples.npy")